[module] Verplaats eetplan naar een module
