+++ htdocs/.htaccess
@@ -23,7 +23,7 @@ RewriteRule ^verify				index.php?c=Login [L]
 RewriteRule ^wachtwoord			index.php?c=Login [L]
 RewriteRule ^account			index.php?c=Login [L]
 
-RewriteRule ^eetplan			index.php?c=Eetplan		[QSA,L]
+RewriteRule ^eetplan			index.php?m=eetplan&c=Eetplan	[QSA,L]
 RewriteRule ^documenten			index.php?c=Documenten	[QSA,L]
 RewriteRule ^bibliotheek		index.php?c=Bibliotheek [QSA,L]
 RewriteRule ^mededelingen       index.php?c=Mededelingen [QSA,L]
