+++ lib/eetplan/model/entity/Eetplan.class.php
